requirements.txt
streamlit==1.38.0
google-genai==0.4.0
python-dotenv